worker/requirements.txt
redis==5.0.1
pymongo==4.5.0
python-dotenv==1.0.0